adds fix to map rotation and movement, also fixes the blocking tile from player position
