+++ maps/valid/easy.cub
@@ -5,7 +5,13 @@ EA ./textures/east.png
 F 250, 0 ,0
 C 250, 0, 250
 
-1111111111
-1000000001
-1E00000001
-1111111111
+1111111111111111
+1000000001111111
+1000000001111111
+1111000111111111
+1111000111111111
+1000000001111111
+1000100001111111
+1000100001111111
+1N00000001111111
+1111111111111111
